+++ Frontend/nubio/src/components/safeHome/route/short/ShortDirection.tsx
@@ -7,7 +7,8 @@ import proj4 from "proj4";
 import Swal from "sweetalert2";
 // 컴포넌트
 // 스타일
-import { ShortDirectionButton } from "../../../../styles/SKakaoMap";
+import { ShortDirectionButton } from "../../../../styles/SRoute";
+
 // redux
 import { setShortTime } from "../../../../redux/slice/SafeSlice";
