+++ Frontend/nubio/src/redux/store.tsx
@@ -3,6 +3,7 @@ import EnjoySlice from "./slice/EnjoySlice";
 import Profileslice from "./slice/Profileslice";
 import MapSlice from "./slice/MapSlice";
 import SignUpSlice from "./slice/SignUpSlice";
+import SafeSlice from "./slice/SafeSlice";
 
 export const store = configureStore({
   reducer: {
@@ -10,5 +11,6 @@ export const store = configureStore({
     profile: Profileslice,
     map: MapSlice,
     signup: SignUpSlice,
+    safe: SafeSlice,
   },
 });
