--- Frontend/nubio/src/styles/STMap.tsx
@@ -1,5 +1,0 @@
-import { styled } from "styled-components";
-
-export const TMapWrapper = styled.div`
-  z-index: 0;
-`;
